Images refactored.
