+++ webapp/src/static/images/images.tsx
@@ -1,0 +1,9 @@
+import logo from './lomap_logo.png'
+import noPicture from './no-pictures-picture.png'
+
+const images: any = {
+    logo: logo,
+    noPicture: noPicture
+}
+
+export default images
